+++ src/ui/Windows/MainEdit.cpp
@@ -1878,6 +1878,25 @@ void DboxMain::OnCopyTwoFactorAuthCode()
   CopyDataToClipBoard(ClipboardDataSource::AuthCode);
 }
 
+void DboxMain::OnCopyCustomFieldValue(UINT nID)
+{
+  ASSERT(nID >= ID_MENUITEM_COPYCUSTOMFIELD1 && nID <= ID_MENUITEM_COPYCUSTOMFIELDMAX);
+
+  const size_t idx = nID - ID_MENUITEM_COPYCUSTOMFIELD1;
+  if (idx >= m_vCustomFieldValues.size())
+    return;
+
+  if (!SelItemOk())
+    return;
+
+  CItemData *pci = getSelectedItem();
+  ASSERT(pci != NULL);
+
+  SetClipboardData(m_vCustomFieldValues[idx]);
+  UpdateLastClipboardAction(ClipboardDataSource::CustomFieldValue);
+  UpdateAccessTime(pci->GetUUID());
+}
+
 void DboxMain::OnDisplayPasswordSubset()
 {
   if (!SelItemOk())