+++ src/ui/Windows/resource2.h
@@ -292,6 +292,9 @@ Do NOT change the value of any resource used in a menu from V3.17 onwards.
 #define ID_MENUITEM_TRAYBROWSEALT1      33501
 #define ID_MENUITEM_TRAYBROWSEALTMAX    33525
 
+#define ID_MENUITEM_COPYCUSTOMFIELD1    33526
+#define ID_MENUITEM_COPYCUSTOMFIELDMAX  33550
+
 // Following for CEditExtn context menu
 // Really string constants, but "logically" they're menus...
 #define IDS_MENUSTRING_UNDO             34100